Fixed ArrayIndexOutOfBoundsException
My app was crashing because I used this library and implemented the ExpanableLinearLayout without any children in it.

I added an if statement to allow my app to run. It may help others too. 

Problem:
If the childCount was equal to 0 than the function onMeasure would try to get the "childSizeList" at index -1 which does not exist.

Thanks for the awesome code, really saved me time overall!